Finish story producer coding work

+++ StoryGPTEntityAPI/Dtos/MetaDataDTO.cs
@@ -5,8 +5,8 @@ namespace StoryGPTEntityAPI.Dtos
 {
     public class MetaDataDTO
     {
-        public int Id { get; set; }
-        public int StoryId { get; set; }
+        public long Id { get; set; }
+        public long StoryId { get; set; }
         public DateTime DateCreated { get; set; }
         public string Author { get; set; } = null!;
         public string Model { get; set; } = null!;
